requirements.txt
fastapi
uvicorn
pydantic
reportlab
gitpython
pytest
androguard
python-multipart
Appium-Python-Client